nix: Set to customized 22.11
Recent projects revealed various regressions in few packages in nix
unstable. So we should remove back to stable release and only update
once a new stable nixpkgs release exists.

On the other hand we need lilypond 2.24 for current mutwo.abjad.
So the solution is to add a nixpkgs fork which is the latest stable
release with our custom patches.